+++ CHANGELOG.md
@@ -6,6 +6,7 @@
     * Change default SSH user from `core` to `centos`
     * Allow changing SSH user without using a proxy
     * Update `dcos node diagnostics` deprecation message
+    * Display version for packages with the same name
 
 ## 2.1-patch.0